LesaNet/requirements.txt
matplotlib
easydict
pyyaml
opencv-python
nibabel
cffi
openpyxl
sklearn
scipy